Move ILogger to IO.Ably.Infrastructure and make it public
