+++ src/IO.Ably.Push.Android/AndroidMobileDevice.cs
@@ -7,6 +7,7 @@ using Android.Content;
 using Android.Gms.Tasks;
 using Android.Support.V4.Content;
 using Firebase.Messaging;
+using IO.Ably.Infrastructure;
 using Xamarin.Essentials;
 
 namespace IO.Ably.Push.Android
