+++ src/IO.Ably.Shared/AblyRealtime.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 using IO.Ably.MessageEncoders;
 using IO.Ably.Push;
 using IO.Ably.Realtime;
