+++ src/IO.Ably.Shared/EventEmitter.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 using Newtonsoft.Json.Linq;
 
 namespace IO.Ably
