+++ src/IO.Ably.Shared/Http/AblyHttpClient.cs
@@ -6,6 +6,7 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
 using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 
 namespace IO.Ably
 {
