+++ src/IO.Ably.Shared/MessageEncoders/MessageHandler.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 using IO.Ably.Realtime;
 using IO.Ably.Types;
