+++ src/IO.Ably.Shared/Push/PushRealtime.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using IO.Ably.Infrastructure;
 
 namespace IO.Ably.Push
 {
