+++ src/IO.Ably.Shared/Realtime/ChannelAwaiter.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 using IO.Ably.Transport;
 using IO.Ably.Transport.States.Connection;
