+++ src/IO.Ably.Shared/Realtime/ChannelMessageProcessor.cs
@@ -1,5 +1,6 @@
 using System.Linq;
 using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 using IO.Ably.MessageEncoders;
 using IO.Ably.Realtime.Workflow;
 using IO.Ably.Types;
