+++ src/IO.Ably.Shared/Realtime/Presence.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 using IO.Ably.Transport;
 using IO.Ably.Types;
 using IO.Ably.Utils;
