+++ src/IO.Ably.Shared/Realtime/RealtimeChannels.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 using IO.Ably.Realtime.Workflow;
 using Newtonsoft.Json.Linq;
