+++ src/IO.Ably.Shared/Realtime/Workflows/RealtimeWorkflow.cs
@@ -6,6 +6,7 @@ using System.Reflection;
 using System.Threading;
 using System.Threading.Channels;
 using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 using IO.Ably.Transport;
 using IO.Ably.Transport.States.Connection;
 using IO.Ably.Types;
