+++ src/IO.Ably.Shared/Rest/ChannelOptions.cs
@@ -1,4 +1,5 @@
 using IO.Ably.Encryption;
+using IO.Ably.Infrastructure;
 
 namespace IO.Ably
 {
