+++ src/IO.Ably.Shared/Transport/ConnectionAttemptsInfo.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using IO.Ably.Infrastructure;
 using IO.Ably.Realtime;
 using IO.Ably.Transport.States.Connection;
