+++ src/IO.Ably.Shared/Transport/MsWebSocketTransport.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 using IO.Ably.Realtime;
 
 namespace IO.Ably.Transport
