+++ src/IO.Ably.Shared/Transport/States/Connection/ConnectionConnectedState.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 using IO.Ably.Realtime;
 using IO.Ably.Realtime.Workflow;
 using IO.Ably.Types;
