+++ src/IO.Ably.Shared/Transport/TransportParams.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Net;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 
 namespace IO.Ably.Transport
 {
