+++ src/IO.Ably.Shared/Utils/ActionUtils.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using IO.Ably.Infrastructure;
 
 namespace IO.Ably.Utils
 {
