+++ src/IO.Ably.Tests.Shared/Rest/SandboxSpec.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
+using IO.Ably.Infrastructure;
 using IO.Ably.Realtime;
 using IO.Ably.Tests.Infrastructure;
 using Xunit;
